Moving dashboard subprocess from pool object to chain object
